Move g2p association code into new net.fnothaft.gnocchi.association package.
