+++ gnocchi-cli/src/main/scala/net/fnothaft/gnocchi/cli/RegressPhenotypes.scala
@@ -17,7 +17,7 @@ package net.fnothaft.gnocchi.cli
 
 import java.io.File
 import net.fnothaft.gnocchi.avro.{ Association, Phenotype }
-import net.fnothaft.gnocchi._
+import net.fnothaft.gnocchi.association._
 import org.apache.hadoop.mapreduce.Job
 import org.apache.spark.SparkContext._
 import org.apache.spark.{ Logging, SparkContext }
