+++ gnocchi-core/src/main/scala/net/fnothaft/gnocchi/association/ScoreAssociation.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.fnothaft.gnocchi
+package net.fnothaft.gnocchi.association
 
 import net.fnothaft.gnocchi.avro.{ Association, Phenotype }
 import org.apache.spark.SparkContext._
@@ -28,7 +28,7 @@ import scala.math.{ log => mathLog }
 private[gnocchi] case class GPKey(allele: String, dosage: Int, hasPhenotype: Boolean) {
 }
 
-object ScoreAssociation extends Serializable with Logging {
+private[gnocchi] object ScoreAssociation extends Serializable with Logging {
 
   // store the natural log of 10 as a constant
   val LOG10 = mathLog(10.0)
@@ -36,7 +36,7 @@ object ScoreAssociation extends Serializable with Logging {
   /**
    *
    */
-  private[gnocchi] def toRecord(kv: (String, (Genotype, Phenotype))): Option[((ReferencePosition, String), GPKey)] = {
+  def toRecord(kv: (String, (Genotype, Phenotype))): Option[((ReferencePosition, String), GPKey)] = {
     // extract the (sample ID, (site genotype, sample phenotype)) from the key-value pair we got
     val (s, (g, p)) = kv
 
@@ -86,7 +86,7 @@ object ScoreAssociation extends Serializable with Logging {
    *         the heterozygous genotype, the odds ratio for an association with the homozygous alt genotype,
    *         the chi squared statistic, and the log probability of the null hypothesis.
    */
-  private[gnocchi] def chiSquared(homRefNP: Int, hetNP: Int, homAltNP: Int,
+  def chiSquared(homRefNP: Int, hetNP: Int, homAltNP: Int,
                                   homRefP: Int, hetP: Int, homAltP: Int): (Double, Double, Double, Double) = {
     // odds ratio for het
     val oHet = (homRefNP * hetP).toDouble / (homRefP * hetNP).toDouble
@@ -128,7 +128,7 @@ object ScoreAssociation extends Serializable with Logging {
    *        we are testing), a map describing the genotypes we've observed).
    * @return If there are any alternate alleles at the site, returns genotype/phenotype association statistics.
    */
-  private[gnocchi] def toAssociation(site: ((ReferencePosition, String), HashMap[GPKey, Int])): Option[Association] = {
+  def toAssociation(site: ((ReferencePosition, String), HashMap[GPKey, Int])): Option[Association] = {
     // unpack observation
     val ((pos, phenotype), map) = site
